Move to watch instead of watcher
