+++ client/src/utils/data.js
@@ -4,7 +4,7 @@ import { FilesDialog } from "components/FilesDialog";
 import { useGlobalUploadModal } from "composables/globalUploadModal";
 import $ from "jquery";
 import { getAppRoot } from "onload/loadConfig";
-import { startWatchingHistory } from "@/watcher/watchHistory";
+import { startWatchingHistory } from "@/watch/watchHistory";
 import Vue from "vue";
 
 import { uploadPayload } from "@/utils/upload-payload.js";
